got both interfaces more or less functional
