+++ dweet2ser/__init__.py
@@ -1,5 +1,2 @@
 __version__ = '0.2.0'
 
-from .session import DweetSession
-
-current_session = DweetSession()
